refactor: initView 메서드 분리
