+++ app/src/main/java/woowacourse/movie/presentation/model/MovieDrawableData.kt
@@ -7,13 +7,13 @@ object MovieDrawableData {
     private val moviePosters = mapOf(
         1L to R.drawable.harry_potter_poster,
         2L to R.drawable.iron_man_poster,
-        3L to R.drawable.suzume_poster
+        3L to R.drawable.suzume_poster,
     )
 
     private val movieThumbnails = mapOf(
         1L to R.drawable.harry_potter_thumbnail,
         2L to R.drawable.iron_man_thumbnail,
-        3L to R.drawable.suzume_thumbnail
+        3L to R.drawable.suzume_thumbnail,
     )
 
     fun getMoviePoster(movieId: Long) = moviePosters[movieId]
